+++ src/sasctl/_services/microanalytic_score.py
@@ -9,6 +9,8 @@
 import re
 from collections import OrderedDict
 
+import six
+
 from .service import Service
 
 
@@ -116,7 +118,18 @@ class MicroAnalyticScore(Service):
         module = module.id
         step = step.id if hasattr(step, 'id') else step
 
-        body = {'inputs': [{'name': k, 'value': v} for k, v in kwargs.items()]}
+        # Make sure all inputs are JSON serializable
+        # Common types such as numpy.int64 and numpy.float64 are NOT serializable
+        for k in kwargs.keys():
+            type_name = type(kwargs[k]).__name__
+            if type_name == 'float64':
+                kwargs[k] = float(kwargs[k])
+            elif type_name == 'int64':
+                kwargs[k] = int(kwargs[k])
+
+
+        body = {'inputs': [{'name': k, 'value': v}
+                           for k, v in six.iteritems(kwargs)]}
         r = self.post('/modules/{}/steps/{}'.format(module, step), json=body)
 
         # Convert list of name/value pair dictionaries to single dict